requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
tabulate==0.9.0
thoughtspot_rest_api_v1==1.8.5
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
